+++ src/services/line/line.service.test.ts
@@ -1,5 +1,5 @@
 import * as sinon from 'sinon';
-import * as Qminder from '../../src/qminder-api';
+import * as Qminder from '../../qminder-api';
 
 describe('LineService', function () {
   const LINES = [
