+++ src/services/ticket/ticket.service.ts
@@ -2,12 +2,12 @@ import Ticket, {
   TicketStatus,
   TicketMessage,
   TicketExtra,
-} from '../model/ticket';
-import User from '../model/user';
-import Desk from '../model/desk';
-import Line from '../model/line';
-import ApiBase from '../api-base';
-import { extractId, extractIdToNumber, IdOrObject } from '../util/id-or-object';
+} from '../../model/ticket';
+import User from '../../model/user';
+import Desk from '../../model/desk';
+import Line from '../../model/line';
+import ApiBase from '../api-base/api-base';
+import { extractId, extractIdToNumber, IdOrObject } from '../../util/id-or-object';
 
 /**
  * Represents a collection of search criteria for TicketService.count().
