+++ src/services/user/user.service.ts
@@ -1,9 +1,9 @@
-import ApiBase, { SuccessResponse } from '../api-base';
-import User, { UserRole } from '../model/user';
-import Desk from '../model/desk';
-import Location from '../model/location';
-import Line from '../model/line';
-import { extractId, extractIdToNumber, IdOrObject } from '../util/id-or-object';
+import ApiBase, { SuccessResponse } from '../api-base/api-base';
+import User, { UserRole } from '../../model/user';
+import Desk from '../../model/desk';
+import Location from '../../model/location';
+import Line from '../../model/line';
+import { extractId, extractIdToNumber, IdOrObject } from '../../util/id-or-object';
 
 /**
  * User Service
