+++ src/services/webhooks/webhooks.service.ts
@@ -1,6 +1,6 @@
-import ApiBase, { SuccessResponse } from '../api-base';
-import Webhook from '../model/webhook';
-import { extractId, IdOrObject } from '../util/id-or-object';
+import ApiBase, { SuccessResponse } from '../api-base/api-base';
+import Webhook from '../../model/webhook';
+import { extractId, IdOrObject } from '../../util/id-or-object';
 
 /** @hidden */
 export const ERROR_NO_URL = 'No URL provided';
